Responsive Design - Phase 1 Add-On 1
Extra functionality is added which allows the user to click on the
"Menu" button to expand the menu. However, the area where the links
should appear is blank. This may be due to the use of the dark shade of
green as both the font color and the bottom-line color. As of right now,
the current navigation bar is just a prototype.